// src/app/drivermaster/models/driver.model.ts
export interface Driver {
  id: number;
  FirstName: string;
  MiddleName: string;
  LastName: string;
  SAPId?: string;
  UserId?: number;
  BranchCode?: string;
  DriverCode: string;
  Location: string;
  MobileNumber: number;
  PermanentAddress: string;
  CurrentAddress: string;
  LicenseNumber: string;
  LicenseValidity: Date;
  IssuedByRTO: string;
  FirstLicenseIssueDate: Date;
  CloseTrip: number;
  DriverFatherName: string;
  VehicleNumber: string;
  PermanentCity: string;
  PermanentPincode: number;
  CurrentCity: string;
  CurrentPincode: number;
  GuarantorName: string;
  Status: "ACTIVE" | "DEACTIVATED" | "DELETED";
  DriverCategory: string;
  DOB: Date;
  DOJ: Date;
  Ethinicity: string;
  CurrentLicenseIssueDate: Date;
  LicenseVerifiedDate: Date;
  LicenseVerified: number;
  VerifiedByUserId: number;
  AddressVerified: number;
  DriverPhoto: string;
  PanCard: string;
  VoterId: string;
  AadharCard: string;
  License: string;
  created_at: Date;
  updated_at: Date;
}